add mkdirp code as few lines of code, remove mkdirp as dependencies
